JKRAram: back ARAM with host memory + real DMA copy (2D archive load works)
The native ARAM seam (ar_dsp_impl.cpp) only allocated addresses; ARQ DMA was a
no-op completion that moved no bytes, so ARAM was always empty. That was harmless
while the only ARAM user was audio (native_jas bypasses it), but the 2D/UI ARCHIVE
path genuinely DMAs through ARAM: TApplication::setupThreadFuncLogo streams
guide.szs/game_6.szs into ARAM (gArBkGuide/gArBkConsole), and SMSSwitch2DArchive
copies them back out (Yaz0-decompressing ARAM -> main RAM) to mount the "guide" 2D
volume. With ARAM empty, JKRFileLoader::getVolume("guide") was null and
TConsoleStr::load OSPanic'd (the stage-15/file-select setupObjects blocker).

Make ARAM real:
- ar_dsp_impl.cpp: ARInit calloc's a 16 MB host buffer; ARQPostRequest/ARStartDMA
  do a direction-aware memcpy between host main-RAM pointers and ARAM offsets,
  fail-fast (OSPanic) on an out-of-range ARAM offset.
- LP64: a main-memory address is a 64-bit host pointer that cannot round-trip
  through the SDK's u32 DMA args (it truncated to a wild pointer). Widen the whole
  ARAM DMA address chain to pointer width (new ARMemAddr = uintptr_t native / u32 on
  GC, mirroring the existing ARQRequestRef): JKRAMCommand mSrc/mDst, JKRAramPcs /
  orderSync / orderAsync / prepareCommand, ARQPostRequest source/dest, ARStartDMA
  mainmem_addr, and drop the truncating (u32)(uintptr_t) casts at every call site
  (JKRAram, JKRAramStream, JKRDvdAramRipper, JASDvdThread). On GC uintptr_t == u32 so
  the on-target layout/behavior is unchanged.

Verified: ar_dsp round-trips bytes through ARAM (render/ctest), and the live boot
mounts "guide" then game_6 (correct decompressed RARC + sizes; mountFixed ok),
clearing the TConsoleStr panic. SB_ARC_DBG dumps the ARAM header + mount result.

+++ include/JSystem/JKernel/JKRAramPiece.hpp
@@ -22,8 +22,11 @@ public:
 
 	/* 0x40 */ s32 mTransferDirection;
 	/* 0x44 */ u32 mDataLength;
-	/* 0x48 */ u32 mSrc;
-	/* 0x4C */ u32 mDst;
+	// Main-memory side carries a host pointer on a 64-bit host; widen to pointer
+	// width (== u32 on GC, so the on-target layout is unchanged). The ARAM side is a
+	// small offset that fits either way. See ARMemAddr in dolphin/ar.h.
+	/* 0x48 */ uintptr_t mSrc;
+	/* 0x50 */ uintptr_t mDst;
 	/* 0x50 */ JKRAramBlock* mAramBlock;
 	/* 0x54 */ u32 field_0x54;
 	/* 0x58 */ AsyncCallback mCallback;
@@ -50,14 +53,15 @@ public:
 	};
 
 public:
-	static JKRAMCommand* prepareCommand(int, u32, u32, u32, JKRAramBlock*,
+	static JKRAMCommand* prepareCommand(int, uintptr_t, uintptr_t, u32,
+	                                    JKRAramBlock*,
 	                                    JKRAMCommand::AsyncCallback);
 	static void sendCommand(JKRAMCommand*);
 
-	static JKRAMCommand* orderAsync(int, u32, u32, u32, JKRAramBlock*,
+	static JKRAMCommand* orderAsync(int, uintptr_t, uintptr_t, u32, JKRAramBlock*,
 	                                JKRAMCommand::AsyncCallback);
 	static bool sync(JKRAMCommand*, int);
-	static bool orderSync(int, u32, u32, u32, JKRAramBlock*);
+	static bool orderSync(int, uintptr_t, uintptr_t, u32, JKRAramBlock*);
 	static void startDMA(JKRAMCommand*);
 	static void doneDMA(ARQRequestRef);
 
@@ -66,8 +70,8 @@ private:
 	static void unlock() { OSUnlockMutex(&mMutex); }
 };
 
-inline BOOL JKRAramPcs(int direction, u32 source, u32 destination, u32 length,
-                       JKRAramBlock* block)
+inline BOOL JKRAramPcs(int direction, uintptr_t source, uintptr_t destination,
+                       u32 length, JKRAramBlock* block)
 {
 	return JKRAramPiece::orderSync(direction, source, destination, length,
 	                               block);
